interfaceと実装のファイルを分離
